Refactor test runner into per-test modules with logging
