+++ tests/hybrid_convection_wave/validate.py
@@ -1,0 +1,17 @@
+#!/usr/bin/env python3
+"""Validator for the hybrid convection-wave test (tests/hybrid_convection_wave).
+
+Uses the shared hybrid-family energy-log validation, plus the rigid-advection
+phase check for the transverse wave.
+"""
+from .._shared.hybrid import validate_hybrid, validate_plot as _hyb_plot
+
+
+def validate_log(pic_diags=None, test_name=None):
+    """Hybrid-wave energy-log validation (shared with the hybrid family)."""
+    return validate_hybrid(pic_diags=pic_diags, test_name=test_name)
+
+
+def validate_plot(test_name):
+    """Convection-wave plot check: rigid advection of the transverse wave."""
+    return _hyb_plot(test_name)
