+++ tests/singlecell/validate.py
@@ -1,0 +1,68 @@
+#!/usr/bin/env python3
+"""Validator for the single-cell periodic hybrid test (tests/singlecell).
+
+With exactly one grid cell and periodic boundaries, curl B is identically zero,
+so the Hall term (J x B)/rho and the convective term U_i x B both vanish
+(U_i = 0).  The electric field stays zero and the magnetic field is frozen.
+"""
+import logging
+import math
+
+logger = logging.getLogger(__name__)
+
+
+def validate_log(pic_diags=None, test_name=None):
+    """Validate the single-cell hybrid test.
+
+    The test passes iff (1) the magnetic energy Eb is conserved to round-off
+    (no spurious Hall-driven evolution), (2) the electric field energy Ee stays
+    ~0 (the field is truly frozen, not merely energy-conserving), and (3) no
+    NaN/blow-up occurs.
+    """
+    logger.debug("=== Validating Single-Cell Hybrid Test ===")
+    if not pic_diags:
+        return False, "No diagnostics found"
+    first = pic_diags[0]
+    last = pic_diags[-1]
+    passed = True
+    reasons = []
+
+    eb0 = first.get("Eb", 0.0)
+    eb1 = last.get("Eb", 0.0)
+    logger.debug("    Eb (magnetic): %s -> %s", f"{eb0:.6e}", f"{eb1:.6e}")
+    if not math.isfinite(eb1):
+        passed = False
+        reasons.append("Eb not finite (NaN/Inf)")
+    if eb0 > 0:
+        ratio = eb1 / eb0
+        logger.debug("    Eb ratio: %.6f", ratio)
+        # Exact (to round-off): a single cell has no spatial gradient, so the
+        # Hall term is exactly zero and B cannot evolve.  Allow a tiny tolerance
+        # for floating-point round-off in the field solvers.
+        if ratio < 0.9999 or ratio > 1.0001:
+            passed = False
+            reasons.append(
+                f"Eb ratio {ratio:.6f} not ~1 (spurious Hall/evolution on "
+                f"single cell; curl B must be zero)")
+
+    # Ee must stay ~0: a frozen field has E = 0, so there is no electric energy.
+    # This is the real discriminator between "frozen" and "propagating
+    # non-dispersively" (both conserve Eb, but only a frozen field has Ee ~ 0).
+    # Allow a small round-off floor: on a single cell the field solver leaves a
+    # residual Ee of order 1e-3 * Eb (vs. ~1 for a genuinely propagating wave),
+    # so a 1e-2 * Eb threshold cleanly separates frozen from evolving.
+    eemax = max((d.get("Ee", 0.0) for d in pic_diags), default=0.0)
+    logger.debug("    Ee (electric, max): %s", f"{eemax:.6e}")
+    if not math.isfinite(eemax):
+        passed = False
+        reasons.append("Ee not finite (NaN/Inf)")
+    if eb0 > 0 and eemax > 1.0e-2 * eb0:
+        passed = False
+        reasons.append(
+            f"Ee {eemax:.3e} not ~0 vs Eb {eb0:.3e} (field is evolving / "
+            f"propagating, not frozen)")
+
+    if passed:
+        logger.info("Single-Cell Hybrid Test: PASSED")
+        return True, "Passed"
+    return False, "; ".join(reasons)
